requirements.txt
xpander-sdk
jupyter
PyYAML==6.0.2
colorama==0.4.6
streamlit==1.39.0
fastapi
uvicorn
boto3
openai
anthropic
python-dotenv